# api/openapi.yaml
openapi: 3.0.3
info:
  title: Shop Public API
  version: 1.4.0
  description: |
    Public and back-office API of the demo shop platform: catalog, carts, orders,
    payments, shipments, reviews, promo codes and the user back office.

    The document is maintained by hand and is the contract of record for the
    `shop-api-spring` service. It is intentionally NOT generated from the running
    application.
  contact:
    name: Shop Platform Team
    email: platform@shop.example.com
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080/shop
    description: Local developer instance (server.servlet.context-path = /shop)
  - url: https://api.shop.example.com/shop
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/shipments:
    get:
      operationId: listShipments
      summary: List shipments
      tags: [shipments]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ShipmentStatus'
        - name: carrier
          in: query
          required: false
          schema:
            type: string
            enum: [CDEK, POCHTA, DHL, BOXBERRY]
      responses:
        '200':
          description: Page of shipments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShipmentPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: createShipment
      summary: Create a shipment for an order
      tags: [shipments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShipmentCreateRequest'
      responses:
        '201':
          description: Shipment created
          headers:
            Location:
              description: URI of the created shipment
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shipment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/shipments/{id}:
    get:
      operationId: getShipment
      summary: Get a single shipment
      tags: [shipments]
      parameters:
        - $ref: '#/components/parameters/ShipmentIdParam'
      responses:
        '200':
          description: Shipment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shipment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/shipments/{id}/documents:
    post:
      operationId: uploadShipmentDocument
      summary: Attach a scanned document to a shipment
      tags: [shipments]
      parameters:
        - $ref: '#/components/parameters/ShipmentIdParam'
        - name: kind
          in: query
          required: true
          description: Document kind
          schema:
            type: string
            enum: [WAYBILL, INVOICE, CUSTOMS]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Document payload
      responses:
        '201':
          description: Document attached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShipmentDocument'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Customer or operator access token issued by the shop identity provider
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Static back-office key, required for every /admin operation
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      description: Zero based page index
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      required: false
      description: Page size
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 200
        default: 20
...
    ShipmentIdParam:
      name: id
      in: path
      required: true
      description: Shipment identifier
      schema:
        type: string
        maxLength: 40
...
  responses:
    BadRequest:
      description: The request payload or parameters failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Unexpected server failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, timestamp]
      properties:
        code:
          type: string
          example: VALIDATION_FAILED
        message:
          type: string
        path:
          type: string
        timestamp:
          type: string
          format: date-time
        details:
          type: array
          items:
            type: string
    PageMeta:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    Address:
      type: object
      required: [country, city, street, postalCode]
      properties:
        country:
          type: string
          pattern: '^[A-Z]{2}$'
        city:
          type: string
          maxLength: 80
        street:
          type: string
          maxLength: 160
        postalCode:
          type: string
          pattern: '^[0-9]{6}$'
...
    ShipmentStatus:
      type: string
      enum: [CREATED, HANDED_OVER, IN_TRANSIT, DELIVERED, RETURNED]
    ShipmentDocument:
      type: object
      required: [id, kind, fileName, sizeBytes]
      properties:
        id:
          type: string
        kind:
          type: string
          enum: [WAYBILL, INVOICE, CUSTOMS]
        fileName:
          type: string
        sizeBytes:
          type: integer
          format: int64
        uploadedAt:
          type: string
          format: date-time
    Shipment:
      type: object
      required: [id, orderId, carrier, status, destination]
      properties:
        id:
          type: string
        orderId:
          type: string
        carrier:
          type: string
          enum: [CDEK, POCHTA, DHL, BOXBERRY]
        trackingNumber:
          type: string
        status:
          $ref: '#/components/schemas/ShipmentStatus'
        destination:
          $ref: '#/components/schemas/Address'
        estimatedDelivery:
          type: string
          format: date
        documents:
          type: array
          description: Nested collection of attached shipping documents
          items:
            $ref: '#/components/schemas/ShipmentDocument'
        createdAt:
          type: string
          format: date-time
    ShipmentCreateRequest:
      type: object
      required: [orderId, carrier, destination, estimatedDelivery]
      properties:
        orderId:
          type: string
          maxLength: 40
        carrier:
          type: string
          enum: [CDEK, POCHTA, DHL, BOXBERRY]
        destination:
          $ref: '#/components/schemas/Address'
        estimatedDelivery:
          type: string
          format: date
    ShipmentPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [content]
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Shipment'
...